TON of fixes, added holesnap

+++ src/main/java/me/lyric/infinity/api/util/minecraft/EntityUtil.java
@@ -14,7 +14,7 @@ public class EntityUtil implements IGlobals {
     public static int getPing(EntityPlayer p) {
         int ping = 0;
         try {
-            ping = (int) mc.getConnection().getPlayerInfo(p.getUniqueID()).getResponseTime();
+            ping = mc.getConnection().getPlayerInfo(p.getUniqueID()).getResponseTime();
         } catch (NullPointerException np) {}
         return ping;
     }
